Extract functions to dedicated objects in packages
