+++ src/main/scala/com/ocadotechnology/Main.scala
@@ -2,41 +2,18 @@ package com.ocadotechnology
 
 import cats.effect.{ExitCode, IO, IOApp}
 import com.comcast.ip4s.{Host, Port, port}
+import com.ocadotechnology.Router.routes
+import com.ocadotechnology.models.User
 import org.http4s.ember.server.EmberServerBuilder
 import org.http4s.server.Router
 import sttp.tapir.server.http4s.Http4sServerInterpreter
-import doobie.*
-import doobie.implicits.*
-import doobie.util.transactor.Transactor.Aux
 
 
 
 object Main extends IOApp:
-
-  val xa: Aux[IO, Unit] = Transactor.fromDriverManager[IO](
-    "org.postgresql.Driver",
-    "jdbc:postgresql://localhost:5432/Doubloons",
-    "postgres",
-    "admin"
-  )
-  def getUser(email: String): IO[Either[String, User]] = {
-    sql"""SELECT * FROM "Users" WHERE "Email" = $email """
-//    sql"""SELECT * FROM "Users" LIMIT 1"""
-      .query[User]
-      .option
-      .transact(xa)
-      .map {
-        case Some(user) => Right(user)
-        case None => Left("User not found")
-      }
-  }
-
-
+  
   override def run(args: List[String]): IO[ExitCode] =
-
-
-    val routes = Http4sServerInterpreter[IO]().toRoutes(Endpoints.all)
-
+    
     val port = sys.env
       .get("HTTP_PORT")
       .flatMap(_.toIntOption)
