+++ src/main/scala/com/ocadotechnology/services/UserService.scala
@@ -1,0 +1,20 @@
+package com.ocadotechnology.services
+
+import cats.effect.IO
+import com.ocadotechnology.models.User
+import com.ocadotechnology.repositories.UserRepository.getUserByEmail
+
+// Business Logic for User model
+object UserService {
+  
+  def getUserByEmailLogic(email: String): IO[Either[String, User]] = {
+    getUserByEmail(email)
+      .map {
+        case Some(user) => Right(user)
+        case None => Left("User not found")
+      }
+  }
+  
+}
+
+
