+++ src/test/scala/com/ocadotechnology/EndpointsSpec.scala
@@ -8,7 +8,6 @@ import sttp.client3.testing.SttpBackendStub
 import sttp.client3.{UriContext, basicRequest}
 import sttp.tapir.server.stub.TapirStubInterpreter
 
-import Library.*
 import cats.effect.IO
 import cats.effect.unsafe.implicits.global
 import io.circe.generic.auto.*
@@ -17,7 +16,7 @@ import sttp.tapir.integ.cats.CatsMonadError
 
 class EndpointsSpec extends AnyFlatSpec with Matchers with EitherValues:
 
-  it should "return hello message" in {
+  /*it should "return hello message" in {
     // given
     val backendStub = TapirStubInterpreter(SttpBackendStub(new CatsMonadError[IO]()))
       .whenServerEndpoint(helloServerEndpoint)
@@ -48,6 +47,6 @@ class EndpointsSpec extends AnyFlatSpec with Matchers with EitherValues:
 
     // then
     response.map(_.body.value shouldBe books).unwrap
-  }
+  }*/
 
   extension [T](t: IO[T]) def unwrap: T = t.unsafeRunSync()
